Implemented std::(w)string insertion and extraction for the Serializer. Serializer.h has to include Definitions.h to get the correct setting of DDS_HAS_WCHAR but this lead to a cyclic include, moved SequenceNumber out of Definitions.h to break this. Updated some other files to include the new SequenceNumber.h. This implements issue #672
    * dds/DCPS/SequenceNumber.h:
      Added.

    * dds/DCPS/CoherentChangeControl.h:
    * dds/DCPS/DataSampleHeader.h:
    * dds/DCPS/Definitions.h:
    * dds/DCPS/DisjointSequence.h:
    * dds/DCPS/Serializer.h:
    * dds/DCPS/Serializer.inl:
    * dds/DCPS/transport/framework/TransportHeader.h:
    * dds/DCPS/transport/framework/TransportQueueElement.h:
    * dds/DCPS/transport/rtps_udp/RtpsTransportHeader.h:
    * tests/DCPS/Serializer/SerializerTest.cpp:
