+++ dds/DCPS/DataSampleHeader.h
@@ -11,6 +11,7 @@
 #include "Definitions.h"
 #include "GuidUtils.h"
 #include "PoolAllocationBase.h"
+#include "SequenceNumber.h"
 
 #include <iosfwd>
