+++ dds/DCPS/transport/framework/TransportQueueElement.h
@@ -12,6 +12,7 @@
 #include "dds/DCPS/Definitions.h"
 #include "dds/DCPS/GuidUtils.h"
 #include "dds/DCPS/PoolAllocationBase.h"
+#include "dds/DCPS/SequenceNumber.h"
 
 #include <utility>
