+++ dds/DCPS/transport/rtps_udp/RtpsTransportHeader.h
@@ -9,6 +9,7 @@
 #define DCPS_RTPSTRANSPORTHEADER_H
 
 #include "dds/DCPS/Definitions.h"
+#include "dds/DCPS/SequenceNumber.h"
 #include "dds/DCPS/RTPS/RtpsCoreC.h"
 
 ACE_BEGIN_VERSIONED_NAMESPACE_DECL
